feat(166-03): extract admin, system, smelter routers and clean main.py
- Removed duplicate job definitions endpoints from main.py (now in jobs_router)
- Verified all 7 CE routers (auth, jobs, nodes, workflows, admin, system, smelter) wired via app.include_router()
- Fixed import errors in admin_router (pki_service path), system_router (AsyncSessionLocal, LicenceState locations), smelter_router (require_permission path)
- All routers now import successfully with no circular dependencies
- main.py contains only legitimate infrastructure routes: compose generators, installers, docs, job templates, retention config, smelter discovery
