+++ puppeteer/agent_service/main.py
@@ -516,44 +516,18 @@ from .routers.auth_router import router as auth_router
 from .routers.jobs_router import router as jobs_router
 from .routers.nodes_router import router as nodes_router
 from .routers.workflows_router import router as workflows_router
+from .routers.admin_router import router as admin_router
+from .routers.system_router import router as system_router
+from .routers.smelter_router import router as smelter_router
 
 # Include CE routers (after middleware setup, before route definitions)
 app.include_router(auth_router, tags=["Authentication"])
 app.include_router(jobs_router, tags=["Jobs", "Job Definitions", "Job Templates", "CI/CD Dispatch"])
 app.include_router(nodes_router, tags=["Nodes", "Node Agent"])
 app.include_router(workflows_router, tags=["Workflows"])
-
-# --- ALERTS ---
-
-@app.get("/api/alerts", response_model=List[AlertResponse], tags=["Alerts & Webhooks"])
-async def list_alerts(
-    skip: int = 0,
-    limit: int = 50,
-    unacknowledged_only: bool = False,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db)
-):
-    """List system alerts with optional filtering."""
-    return await AlertService.list_alerts(db, skip, limit, unacknowledged_only)
-
-@app.post(
-    "/api/alerts/{alert_id}/acknowledge",
-    response_model=ActionResponse,
-    tags=["Alerts & Webhooks"],
-    summary="Acknowledge an alert",
-    description="Mark an alert as acknowledged by ID"
-)
-async def acknowledge_alert(
-    alert_id: int,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db)
-):
-    """Mark an alert as acknowledged."""
-    alert = await AlertService.acknowledge_alert(db, alert_id, current_user.username)
-    if not alert:
-        raise HTTPException(status_code=404, detail="Alert not found")
-    await db.commit()
-    return {"status": "acknowledged", "resource_type": "alert", "resource_id": alert_id}
+app.include_router(admin_router, tags=["Admin", "Signatures", "Alerts & Webhooks"])
+app.include_router(system_router, tags=["System", "Health", "Schedule"])
+app.include_router(smelter_router, tags=["Foundry", "Blueprints"])
 
 # Serve Installer Scripts
 @app.get(
@@ -876,1387 +850,6 @@ class ConnectionManager:
 
 ws_manager = ConnectionManager()
 
-# --- Auth Endpoints ---
-
-# --- OAuth Device Flow (RFC 8628) ---
-_device_codes: dict[str, dict] = {}
-_user_code_index: dict[str, str] = {}  # user_code -> device_code (reverse index)
-_USER_CODE_ALPHABET = "ABCDEFGHJKMNPQRSTUVWXYZ23456789"  # excludes 0,O,1,I,L
-_DEVICE_TTL_SECONDS = 300   # 5 minutes
-_POLL_INTERVAL_SECONDS = 5
-
-def _generate_user_code() -> str:
-    p1 = "".join(_secrets.choice(_USER_CODE_ALPHABET) for _ in range(4))
-    p2 = "".join(_secrets.choice(_USER_CODE_ALPHABET) for _ in range(4))
-    return f"{p1}-{p2}"
-
-@app.post("/auth/device", response_model=DeviceCodeResponse, tags=["Authentication"])
-async def device_authorization():
-    """RFC 8628 Device Authorization Request — issues device_code and user_code."""
-    now = datetime.utcnow()
-    # Lazy cleanup: evict expired entries (2x TTL = 10 min grace)
-    expired_keys = [k for k, v in list(_device_codes.items()) if v["expiry"] < now]
-    for k in expired_keys:
-        uc = _device_codes.pop(k, {}).get("user_code")
-        if uc:
-            _user_code_index.pop(uc, None)
-
-    device_code = _secrets.token_urlsafe(32)
-    user_code = _generate_user_code()
-    expiry = now + timedelta(seconds=_DEVICE_TTL_SECONDS)
-
-    _device_codes[device_code] = {
-        "user_code": user_code,
-        "expiry": expiry,
-        "status": "pending",
-        "approved_by": None,
-        "last_poll": None,
-    }
-    _user_code_index[user_code] = device_code
-
-    agent_url = os.getenv("AGENT_URL", "https://localhost:8001")
-    return {
-        "device_code": device_code,
-        "user_code": user_code,
-        "verification_uri": f"{agent_url}/auth/device/approve",
-        "verification_uri_complete": f"{agent_url}/auth/device/approve?user_code={user_code}",
-        "expires_in": _DEVICE_TTL_SECONDS,
-        "interval": _POLL_INTERVAL_SECONDS,
-    }
-
-class DeviceTokenRequest(BaseModel):
-    device_code: str
-    grant_type: str = "urn:ietf:params:oauth:grant-type:device_code"
-
-@app.post("/auth/device/token", response_model=TokenResponse, tags=["Authentication"])
-async def device_token_exchange(req: DeviceTokenRequest, db: AsyncSession = Depends(get_db)):
-    """RFC 8628 Device Access Token Request — exchange device_code for JWT."""
-    entry = _device_codes.get(req.device_code)
-    now = datetime.utcnow()
-
-    if not entry:
-        raise HTTPException(400, detail={"error": "expired_token"})
-    if entry["expiry"] < now:
-        uc = _device_codes.pop(req.device_code, {}).get("user_code")
-        if uc:
-            _user_code_index.pop(uc, None)
-        raise HTTPException(400, detail={"error": "expired_token"})
-    if entry["status"] == "denied":
-        raise HTTPException(400, detail={"error": "access_denied"})
-
-    # RFC 8628 slow_down: if polled again before interval
-    last_poll = entry.get("last_poll")
-    if last_poll and (now - last_poll).total_seconds() < _POLL_INTERVAL_SECONDS:
-        entry["last_poll"] = now
-        raise HTTPException(400, detail={"error": "slow_down"})
-    entry["last_poll"] = now
-
-    if entry["status"] == "pending":
-        raise HTTPException(400, detail={"error": "authorization_pending"})
-
-    # status == "approved"
-    username = entry["approved_by"]
-    result = await db.execute(select(User).where(User.username == username))
-    user = result.scalar_one_or_none()
-    if not user:
-        raise HTTPException(400, detail={"error": "access_denied"})
-
-    token = create_access_token(
-        data={"sub": user.username, "tv": user.token_version, "type": "device_flow"},
-        expires_delta=timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES),
-    )
-    # Consume code — one-time use
-    _device_codes.pop(req.device_code, None)
-    _user_code_index.pop(entry["user_code"], None)
-
-    audit(db, user, "device_flow:token_issued", None, {"username": user.username})
-    await db.commit()
-    return TokenResponse(access_token=token, token_type="bearer", must_change_password=user.must_change_password)
-
-@app.get("/auth/device/approve", response_class=HTMLResponse, tags=["Authentication"])
-async def device_approve_page(user_code: str = ""):
-    """Serve the device authorization approval page (inline HTML, no build step)."""
-    # SEC-01: escape user_code before inserting into HTML to prevent XSS
-    escaped_code = _html.escape(user_code or "")
-    return HTMLResponse(content=f"""<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Authorize Device — Master of Puppets</title>
-  <style>
-    body {{ font-family: system-ui, sans-serif; max-width: 480px; margin: 80px auto; padding: 0 1rem; color: #1a1a1a; }}
-    .card {{ background: #f8f9fa; border: 1px solid #dee2e6; border-radius: 8px; padding: 2rem; text-align: center; }}
-    .code {{ font-family: monospace; font-size: 2rem; font-weight: bold; letter-spacing: 0.2em; color: #0d6efd; margin: 1rem 0; }}
-    .btn {{ display: inline-block; padding: 0.6rem 1.6rem; border: none; border-radius: 6px; font-size: 1rem; cursor: pointer; margin: 0.3rem; }}
-    .btn-approve {{ background: #198754; color: white; }}
-    .btn-deny {{ background: #dc3545; color: white; }}
-    .btn-approve:hover {{ background: #157347; }}
-    .btn-deny:hover {{ background: #bb2d3b; }}
-    .msg {{ margin-top: 1rem; font-size: 0.9rem; color: #6c757d; }}
-  </style>
-</head>
-<body>
-  <div class="card">
-    <h2>Authorize Device</h2>
-    <p>A CLI device is requesting access to <strong>Master of Puppets</strong>.</p>
-    <p>Confirm that your terminal displays this code:</p>
-    <div class="code" id="display-code">{escaped_code or "(no code provided)"}</div>
-    <form id="approve-form" method="POST" action="/auth/device/approve">
-      <input type="hidden" name="user_code" value="{escaped_code}">
-      <input type="hidden" name="token" id="token-field" value="">
-      <button type="submit" class="btn btn-approve">Approve</button>
-    </form>
-    <form id="deny-form" method="POST" action="/auth/device/deny">
-      <input type="hidden" name="user_code" value="{escaped_code}">
-      <input type="hidden" name="token" id="deny-token-field" value="">
-      <button type="submit" class="btn btn-deny">Deny</button>
-    </form>
-    <p class="msg" id="auth-msg"></p>
-  </div>
-  <script>
-    document.addEventListener('DOMContentLoaded', function() {{
-      var token = localStorage.getItem('access_token') || '';
-      document.getElementById('token-field').value = token;
-      document.getElementById('deny-token-field').value = token;
-      if (!token) {{
-        document.getElementById('auth-msg').textContent = 'You must be logged in to authorize a device.';
-        document.getElementById('auth-msg').style.color = '#dc3545';
-        var next = encodeURIComponent(window.location.href);
-        setTimeout(function() {{ window.location.href = '/login?next=' + next; }}, 2000);
-      }}
-    }});
-  </script>
-</body>
-</html>""")
-
-@app.post("/auth/device/approve", response_class=HTMLResponse, tags=["Authentication"])
-async def device_approve_submit(
-    user_code: str = Form(...),
-    token: str = Form(""),
-    db: AsyncSession = Depends(get_db),
-):
-    """Process device approval — sets status='approved' on matching device code."""
-    # Validate the user's JWT from the form
-    try:
-        payload = verify_token(token)
-        username = payload.get("sub")
-    except Exception:
-        return HTMLResponse(content="<h2>Error: Invalid or missing session token. Please log in and try again.</h2>", status_code=401)
-
-    device_code = _user_code_index.get(user_code)
-    if not device_code or device_code not in _device_codes:
-        return HTMLResponse(content="<h2>Error: Device code not found or expired.</h2>", status_code=404)
-
-    entry = _device_codes[device_code]
-    if entry["expiry"] < datetime.utcnow():
-        return HTMLResponse(content="<h2>Error: Device code has expired.</h2>", status_code=410)
-
-    entry["status"] = "approved"
-    entry["approved_by"] = username
-
-    result = await db.execute(select(User).where(User.username == username))
-    user = result.scalar_one_or_none()
-    if user:
-        audit(db, user, "device_flow:approved", None, {"user_code": user_code})
-        await db.commit()
-
-    return HTMLResponse(content="""<!DOCTYPE html><html><head><title>Authorized</title>
-<style>body{font-family:system-ui,sans-serif;max-width:480px;margin:80px auto;text-align:center;}</style>
-</head><body><h2 style="color:#198754">Device authorized.</h2>
-<p>You may close this tab. Your CLI session is now active.</p></body></html>""")
-
-@app.post("/auth/device/deny", response_class=HTMLResponse, tags=["Authentication"])
-async def device_deny_submit(
-    user_code: str = Form(...),
-    token: str = Form(""),
-    db: AsyncSession = Depends(get_db),
-):
-    """Process device denial — sets status='denied' on matching device code."""
-    try:
-        payload = verify_token(token)
-        username = payload.get("sub")
-    except Exception:
-        username = "unknown"
-
-    device_code = _user_code_index.get(user_code)
-    if device_code and device_code in _device_codes:
-        _device_codes[device_code]["status"] = "denied"
-
-    result = await db.execute(select(User).where(User.username == username))
-    user = result.scalar_one_or_none()
-    if user:
-        audit(db, user, "device_flow:denied", None, {"user_code": user_code})
-        await db.commit()
-
-    return HTMLResponse(content="""<!DOCTYPE html><html><head><title>Denied</title>
-<style>body{font-family:system-ui,sans-serif;max-width:480px;margin:80px auto;text-align:center;}</style>
-</head><body><h2 style="color:#dc3545">Device authorization denied.</h2>
-<p>The CLI request has been rejected. You may close this tab.</p></body></html>""")
-
-@app.post("/auth/login", response_model=TokenResponse, tags=["Authentication"])
-@limiter.limit("5/minute")
-async def login_for_access_token(request: Request, form_data: OAuth2PasswordRequestForm = Depends(), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(User).where(User.username == form_data.username))
-    user = result.scalar_one_or_none()
-    if not user or not verify_password(form_data.password, user.password_hash):
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Incorrect username or password",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    access_token = create_access_token(
-        data={"sub": user.username, "tv": user.token_version, "role": user.role},
-        expires_delta=access_token_expires
-    )
-    return {"access_token": access_token, "token_type": "bearer", "must_change_password": user.must_change_password}
-
-@app.get("/auth/me", response_model=UserResponse, tags=["Authentication"])
-async def read_users_me(current_user: User = Depends(get_current_user)):
-    return UserResponse(id=current_user.username, username=current_user.username, role=current_user.role, created_at=current_user.created_at)
-
-@app.patch("/auth/me", response_model=TokenResponse, tags=["Authentication"])
-async def update_self(req: dict, current_user: User = Depends(get_current_user), db: AsyncSession = Depends(get_db)):
-    """Allow a logged-in user to change their own password.
-    Returns a fresh access token so the current session continues uninterrupted."""
-    new_password = req.get("password", "").strip()
-    if not new_password or len(new_password) < 8:
-        raise HTTPException(status_code=400, detail="Password must be at least 8 characters")
-    # Skip current_password check only when user is in force-change mode (they just authenticated)
-    if not current_user.must_change_password:
-        current_password = req.get("current_password", "")
-        if not current_password or not verify_password(current_password, current_user.password_hash):
-            raise HTTPException(status_code=400, detail="Current password is incorrect")
-    current_user.password_hash = get_password_hash(new_password)
-    current_user.must_change_password = False
-    current_user.token_version = (current_user.token_version or 0) + 1
-    audit(db, current_user, "user:password_changed", detail={"username": current_user.username})
-    await db.commit()
-    # Issue a new token for the current session (old tokens for other sessions are now invalid)
-    new_token = create_access_token(
-        data={"sub": current_user.username, "tv": current_user.token_version, "role": current_user.role},
-        expires_delta=timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    )
-    return TokenResponse(access_token=new_token, token_type="bearer", must_change_password=False)
-
-# --- Core Endpoints ---
-
-@app.get(
-    "/",
-    response_model=dict,
-    tags=["System"],
-    summary="Health check endpoint",
-    description="Simple health check that returns service status and mirror availability"
-)
-async def health_check():
-    mirrors_available = getattr(app.state, "mirrors_available", True)
-    return {
-        "status": "healthy",
-        "service": "Agent Service v0.7",
-        "mirrors_available": mirrors_available
-    }
-
-@app.get("/system/health", response_model=SystemHealthResponse, tags=["System"])
-async def system_health():
-    mirrors_available = getattr(app.state, "mirrors_available", True)
-    return {
-        "status": "healthy",
-        "mirrors_available": mirrors_available
-    }
-
-
-@app.get("/api/health/scheduling", response_model=SchedulingHealthResponse, tags=["Health"])
-async def get_scheduling_health_endpoint(
-    window: str = "24h",
-    current_user: User = Depends(require_permission("jobs:read")),
-    db: AsyncSession = Depends(get_db),
-):
-    if window not in ("24h", "7d", "30d"):
-        raise HTTPException(status_code=422, detail="window must be 24h, 7d, or 30d")
-    data = await scheduler_service.get_scheduling_health(window, db)
-    return SchedulingHealthResponse(**data, window=window)
-
-
-@app.get("/health/scale", response_model=ScaleHealthResponse, tags=["Health"])
-async def get_scale_health_endpoint(
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db),
-):
-    """Return live pool and scheduler health metrics."""
-    from .db import engine, IS_POSTGRES
-    from .services.scheduler_service import scheduler_service
-    from sqlalchemy import func, select as sa_select
-
-    # APScheduler job count
-    apscheduler_jobs = len(scheduler_service.scheduler.get_jobs())
-
-    # Pending job depth
-    result = await db.execute(
-        sa_select(func.count(Job.guid)).where(Job.status == "PENDING")
-    )
-    pending_depth = result.scalar() or 0
-
-    if not IS_POSTGRES:
-        return ScaleHealthResponse(
-            is_postgres=False,
-            pool_size=None,
-            checked_out=None,
-            available=None,
-            overflow=None,
-            apscheduler_jobs=apscheduler_jobs,
-            pending_job_depth=pending_depth,
-        )
-
-    pool = engine.pool
-    return ScaleHealthResponse(
-        is_postgres=True,
-        pool_size=pool.size(),
-        checked_out=pool.checkedout(),
-        available=pool.checkedin(),
-        overflow=pool.overflow(),
-        apscheduler_jobs=apscheduler_jobs,
-        pending_job_depth=pending_depth,
-    )
-
-
-@app.get("/api/features", response_model=FeaturesResponse, tags=["System"])
-async def get_features(request: Request):
-    ctx = getattr(request.app.state, "ee", None)
-    if ctx is None:
-        return {"audit": False, "foundry": False, "webhooks": False,
-                "triggers": False, "rbac": False, "resource_limits": False,
-                "service_principals": False, "api_keys": False,
-                "executions": False}
-    return {
-        "audit": ctx.audit,
-        "foundry": ctx.foundry,
-        "webhooks": ctx.webhooks,
-        "triggers": ctx.triggers,
-        "rbac": ctx.rbac,
-        "resource_limits": ctx.resource_limits,
-        "service_principals": ctx.service_principals,
-        "api_keys": ctx.api_keys,
-        "executions": ctx.executions,
-    }
-
-@app.get("/api/licence", response_model=LicenceStatusResponse, tags=["System"])
-async def get_licence_status(request: Request, current_user: User = Depends(require_auth)):
-    """Returns current licence status. Requires authentication."""
-    ls: Optional[LicenceState] = getattr(request.app.state, "licence_state", None)
-    ee_error = getattr(request.app.state, "ee_activation_error", None)
-    if ls is None:
-        # CE mode — no licence loaded
-        return {
-            "status": "ce",
-            "days_until_expiry": 0,
-            "node_limit": 0,
-            "tier": "ce",
-            "customer_id": None,
-            "grace_days": 0,
-            "ee_activation_error": ee_error,
-        }
-    return {
-        "status": ls.status.value if hasattr(ls.status, "value") else str(ls.status),
-        "days_until_expiry": ls.days_until_expiry,
-        "node_limit": ls.node_limit,
-        "tier": ls.tier,
-        "customer_id": ls.customer_id,
-        "grace_days": ls.grace_days,
-        "ee_activation_error": ee_error,
-    }
-
-
-@app.get(
-    "/jobs",
-    response_model=PaginatedJobResponse,
-    tags=["Jobs"],
-    summary="List all jobs",
-    description="Retrieve a paginated list of all jobs with optional filtering by status, runtime, tags, and other criteria."
-)
-async def list_jobs(
-    cursor: Optional[str] = None,
-    limit: int = 50,
-    status: Optional[str] = None,
-    runtime: Optional[str] = None,
-    task_type: Optional[str] = None,
-    node_id: Optional[str] = None,
-    tags: Optional[str] = None,
-    created_by: Optional[str] = None,
-    date_from: Optional[datetime] = None,
-    date_to: Optional[datetime] = None,
-    search: Optional[str] = None,
-    current_user: User = Depends(require_permission("jobs:read")),
-    db: AsyncSession = Depends(get_db),
-):
-    tags_list = [t.strip() for t in tags.split(",") if t.strip()] if tags else None
-    result = await JobService.list_jobs(
-        db, limit=limit, cursor=cursor,
-        status=status, runtime=runtime, task_type=task_type,
-        node_id=node_id, tags=tags_list,
-        created_by=created_by, date_from=date_from, date_to=date_to, search=search,
-    )
-    return result  # {items, total, next_cursor}
-
-@app.get(
-    "/jobs/count",
-    response_model=JobCountResponse,
-    tags=["Jobs"],
-    summary="Get job count",
-    description="Get total count of jobs, optionally filtered by status."
-)
-async def count_jobs(status: Optional[str] = None, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    from sqlalchemy import func as sqlfunc
-    query = select(sqlfunc.count()).select_from(Job).where(Job.task_type != 'system_heartbeat')
-    if status and status.upper() != 'ALL':
-        query = query.where(Job.status == status.upper())
-    result = await db.execute(query)
-    return {"total": result.scalar()}
-
-@app.get(
-    "/jobs/export",
-    response_class=StreamingResponse,
-    tags=["Jobs"],
-    summary="Export jobs as CSV",
-    description="Export filtered job records as a CSV file with streaming response"
-)
-async def export_jobs(
-    status: Optional[str] = None,
-    runtime: Optional[str] = None,
-    task_type: Optional[str] = None,
-    node_id: Optional[str] = None,
-    tags: Optional[str] = None,
-    created_by: Optional[str] = None,
-    date_from: Optional[datetime] = None,
-    date_to: Optional[datetime] = None,
-    search: Optional[str] = None,
-    current_user: User = Depends(require_permission("jobs:read")),
-    db: AsyncSession = Depends(get_db),
-):
-    tags_list = [t.strip() for t in tags.split(",") if t.strip()] if tags else None
-    EXPORT_LIMIT = 10_000
-    jobs = await JobService.list_jobs_for_export(
-        db, limit=EXPORT_LIMIT,
-        status=status, runtime=runtime, task_type=task_type,
-        node_id=node_id, tags=tags_list,
-        created_by=created_by, date_from=date_from, date_to=date_to, search=search,
-    )
-
-    HEADERS = ["guid", "name", "status", "task_type", "display_type", "runtime",
-               "node_id", "created_at", "started_at", "completed_at", "duration_seconds", "target_tags"]
-
-    def generate():
-        buf = io.StringIO()
-        writer = csv.writer(buf)
-        writer.writerow(HEADERS)
-        yield buf.getvalue()
-        buf.seek(0)
-        buf.truncate()
-        for job in jobs:
-            writer.writerow([
-                job.get("guid", ""), job.get("name", ""), job.get("status", ""),
-                job.get("task_type", ""), job.get("display_type", ""), job.get("runtime", ""),
-                job.get("node_id", ""), job.get("created_at", ""), job.get("started_at", ""),
-                job.get("completed_at", ""), job.get("duration_seconds", ""),
-                ",".join(job.get("target_tags") or []),
-            ])
-            yield buf.getvalue()
-            buf.seek(0)
-            buf.truncate()
-
-    return StreamingResponse(
-        generate(),
-        media_type="text/csv",
-        headers={
-            "Content-Disposition": "attachment; filename=jobs-export.csv",
-            "X-Content-Type-Options": "nosniff",
-        },
-    )
-
-@app.get(
-    "/api/jobs/stats",
-    response_model=JobStatsResponse,
-    tags=["Jobs"],
-    summary="Get job statistics",
-    description="Retrieve aggregated job statistics including counts by status and success rate."
-)
-async def get_job_stats(current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    """Backend Stats for Dashboard charts."""
-    return await JobService.get_job_stats(db)
-
-
-# ---------------------------------------------------------------------------
-# CI/CD Dispatch API  (ENVTAG-04)
-# ---------------------------------------------------------------------------
-
-_TERMINAL_STATUSES = {"COMPLETED", "FAILED", "DEAD_LETTER", "SECURITY_REJECTED"}
-
-
-@app.post("/api/dispatch", response_model=DispatchResponse, tags=["CI/CD Dispatch"])
-async def dispatch_job(
-    req: DispatchRequest,
-    request: Request,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db),
-):
-    """CI/CD dispatch endpoint. Creates a job from a job definition and returns a poll URL.
-    Intended caller: service principals with jobs:write permission.
-    No-node condition: job is created as PENDING; pipelines detect timeout by polling poll_url."""
-
-    # 1. Fetch the job definition
-    result = await db.execute(select(ScheduledJob).where(ScheduledJob.id == req.job_definition_id))
-    s_job = result.scalar_one_or_none()
-    if not s_job:
-        raise HTTPException(
-            status_code=404,
-            detail={"error": "job_definition_not_found", "job_definition_id": req.job_definition_id},
-        )
-
-    # 1b. Reject REVOKED definitions — cannot dispatch a job from a revoked definition
-    if s_job.status == "REVOKED":
-        raise HTTPException(
-            status_code=409,
-            detail={
-                "error": "job_definition_revoked",
-                "id": s_job.id,
-                "message": "Cannot dispatch a REVOKED job definition.",
-            },
-        )
-
-    # 2. Resolve env_tag: dispatch request overrides definition's env_tag; fall back to definition
-    effective_env_tag = req.env_tag if req.env_tag is not None else s_job.env_tag
-
-    # 3. Build JobCreate
-    runtime = getattr(s_job, 'runtime', None) or 'python'
-    payload_dict = {
-        "script_content": s_job.script_content,
-        "signature": s_job.signature_payload,
-        "secrets": {},
-        "runtime": runtime,
-    }
-    job_create = JobCreate(
-        task_type="script",
-        runtime=runtime,
-        payload=payload_dict,
-        target_tags=json.loads(s_job.target_tags) if s_job.target_tags else None,
-        capability_requirements=json.loads(s_job.capability_requirements) if s_job.capability_requirements else None,
-        scheduled_job_id=s_job.id,
-        max_retries=req.max_retries if req.max_retries is not None else s_job.max_retries,
-        backoff_multiplier=s_job.backoff_multiplier,
-        timeout_minutes=req.timeout_minutes if req.timeout_minutes is not None else s_job.timeout_minutes,
-        env_tag=effective_env_tag,
-    )
-
-    # 4. Create the job
-    job_result = await JobService.create_job(job_create, db)
-    job_guid = job_result["guid"]
-
-    # 5. Build poll_url — use PUBLIC_URL env var to avoid localhost in Docker
-    public_url = os.getenv("PUBLIC_URL", str(request.base_url).rstrip("/"))
-    poll_url = f"{public_url}/api/dispatch/{job_guid}/status"
-
-    # 6. Audit (sync — do not await, must be before db.commit)
-    audit(db, current_user, "dispatch_job", job_guid,
-          {"job_definition_id": req.job_definition_id, "env_tag": effective_env_tag})
-    await db.commit()
-
-    return DispatchResponse(
-        job_guid=job_guid,
-        status=job_result.get("status", "PENDING"),
-        job_definition_id=s_job.id,
-        job_definition_name=s_job.name,
-        env_tag=effective_env_tag,
-        poll_url=poll_url,
-    )
-
-
-@app.get("/api/dispatch/{job_guid}/status", response_model=DispatchStatusResponse, tags=["CI/CD Dispatch"])
-async def get_dispatch_status(
-    job_guid: str,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db),
-):
-    """CI/CD poll endpoint. Returns structured status for a dispatched job.
-    Poll this URL until is_terminal=True to detect pass/fail in pipelines."""
-
-    # 1. Fetch job
-    result = await db.execute(select(Job).where(Job.guid == job_guid))
-    job = result.scalar_one_or_none()
-    if not job:
-        raise HTTPException(
-            status_code=404,
-            detail={"error": "job_not_found", "job_guid": job_guid},
-        )
-
-    # 2. Fetch most recent execution record for exit_code
-    er_result = await db.execute(
-        select(ExecutionRecord)
-        .where(ExecutionRecord.job_guid == job_guid)
-        .order_by(ExecutionRecord.completed_at.desc())
-        .limit(1)
-    )
-    latest_record = er_result.scalar_one_or_none()
-
-    return DispatchStatusResponse(
-        job_guid=job.guid,
-        status=job.status,
-        exit_code=latest_record.exit_code if latest_record else None,
-        node_id=job.node_id,
-        attempt=job.retry_count + 1,
-        started_at=job.started_at,
-        completed_at=job.completed_at,
-        is_terminal=job.status in _TERMINAL_STATUSES,
-    )
-
-
-@app.post("/jobs", response_model=JobResponse, tags=["Jobs"])
-async def create_job(job_req: JobCreate, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    try:
-        # SRCH-03: stamp submitter username so Jobs view can filter by creator
-        job_req = job_req.model_copy(update={"created_by": current_user.username})
-
-        # SEC-JOB: If the payload carries a user signature, verify it server-side
-        # against the registered public key, then countersign with the server's own
-        # Ed25519 signing key so the node can verify using its cached verification.key.
-        payload_dict = dict(job_req.payload)
-        user_sig = payload_dict.get("signature")
-        sig_id = payload_dict.get("signature_id")
-        script_content = payload_dict.get("script_content")
-
-        # WIN-05: Normalize CRLF → LF before signature verification and countersigning.
-        # This matches the normalization in node.py (line 585) so both sides agree on bytes.
-        if script_content:
-            script_content = script_content.replace('\r\n', '\n').replace('\r', '\n')
-            payload_dict["script_content"] = script_content
-
-        if user_sig and sig_id and script_content:
-            # 1. Verify user's signature against the registered public key
-            sig_result = await db.execute(select(Signature).where(Signature.id == sig_id))
-            sig_rec = sig_result.scalar_one_or_none()
-            if not sig_rec:
-                raise HTTPException(status_code=422, detail=f"Signature key ID '{sig_id}' not found in registry")
-            try:
-                from .services.signature_service import SignatureService as _SS
-                _SS.verify_payload_signature(sig_rec.public_key, user_sig, script_content)
-            except Exception as _ve:
-                raise HTTPException(status_code=422, detail=f"Signature verification failed: {_ve}")
-
-        # 2. Countersign script with the server's Ed25519 signing key so the
-        #    node can verify using its fetched verification.key (which is the
-        #    server's public counterpart). This is mandatory for all job scripts.
-        if script_content:
-            try:
-                from .services.signature_service import SignatureService
-                server_sig = SignatureService.countersign_for_node(script_content)
-                # Set server countersignature so node verifies correctly
-                payload_dict["signature"] = server_sig
-                job_req = job_req.model_copy(update={"payload": payload_dict})
-            except Exception as e:
-                raise HTTPException(status_code=500, detail="Server signing key unavailable — contact admin")
-
-        result = await JobService.create_job(job_req, db)
-        await ws_manager.broadcast("job:created", {"guid": result["guid"], "status": "PENDING", "task_type": job_req.task_type})
-        return result
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-@app.get("/jobs/{guid}", response_model=JobResponse, tags=["Jobs"])
-async def get_job(guid: str, current_user: User = Depends(require_permission("jobs:read")), db: AsyncSession = Depends(get_db)):
-    """Retrieve a single job by its GUID."""
-    result = await db.execute(select(Job).where(Job.guid == guid))
-    job = result.scalar_one_or_none()
-    if not job:
-        raise HTTPException(status_code=404, detail="Job not found")
-    payload = job.payload if isinstance(job.payload, dict) else json.loads(job.payload or '{}')
-    result_val = job.result if isinstance(job.result, dict) else json.loads(job.result or 'null') if job.result else None
-    target_tags = job.target_tags if isinstance(job.target_tags, list) else json.loads(job.target_tags or 'null') if job.target_tags else None
-
-    # Calculate duration_seconds from started_at and completed_at
-    duration = None
-    if job.started_at and job.completed_at:
-        duration = int((job.completed_at - job.started_at).total_seconds())
-
-    return JobResponse(
-        guid=job.guid,
-        status=job.status,
-        payload=payload,
-        result=result_val,
-        node_id=job.node_id,
-        started_at=job.started_at,
-        duration_seconds=duration,
-        target_tags=target_tags,
-        task_type=job.task_type,
-        display_type=getattr(job, 'display_type', None),
-        name=getattr(job, 'name', None),
-        created_by=getattr(job, 'created_by', None),
-        created_at=job.created_at,
-        runtime=getattr(job, 'runtime', None),
-        originating_guid=getattr(job, 'originating_guid', None),
-    )
-
-@app.patch(
-    "/jobs/{guid}/cancel",
-    response_model=ActionResponse,
-    tags=["Jobs"],
-    summary="Cancel a job",
-    description="Cancel a PENDING or ASSIGNED job, transitioning it to CANCELLED status."
-)
-async def cancel_job(guid: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Job).where(Job.guid == guid))
-    job = result.scalar_one_or_none()
-    if not job:
-        raise HTTPException(status_code=404, detail="Job not found")
-    if job.status not in ("PENDING", "ASSIGNED"):
-        raise HTTPException(status_code=409, detail=f"Cannot cancel a job with status {job.status}")
-    job.status = "CANCELLED"
-    job.completed_at = datetime.utcnow()
-    audit(db, current_user, "job:cancel", guid)
-    await db.commit()
-    await ws_manager.broadcast("job:updated", {"guid": guid, "status": "CANCELLED"})
-    return {"status": "cancelled", "resource_type": "job", "resource_id": guid}
-
-@app.get(
-    "/jobs/{guid}/dispatch-diagnosis",
-    response_model=DispatchDiagnosisResponse,
-    tags=["Jobs"],
-    summary="Get dispatch diagnosis",
-    description="Get diagnostic information explaining why a PENDING job has not yet been dispatched to a node."
-)
-async def get_dispatch_diagnosis(guid: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    """Returns structured explanation for why a PENDING job has not yet dispatched."""
-    result = await JobService.get_dispatch_diagnosis(guid, db)
-    if result.get("reason") == "not_found":
-        raise HTTPException(status_code=404, detail="Job not found")
-    return result
-
-
-@app.post(
-    "/jobs/dispatch-diagnosis/bulk",
-    response_model=BulkDispatchDiagnosisResponse,
-    tags=["Jobs"],
-    summary="Get bulk dispatch diagnosis",
-    description="Get dispatch diagnostic information for multiple jobs in one request."
-)
-async def bulk_dispatch_diagnosis(
-    req: BulkDiagnosisRequest,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db),
-):
-    """Returns dispatch diagnosis for multiple jobs in one call (Phase 88 — DIAG-01)."""
-    results = {}
-    for guid in req.guids:
-        results[guid] = await JobService.get_dispatch_diagnosis(guid, db)
-    return {"results": results}
-
-
-@app.post(
-    "/jobs/{guid}/retry",
-    response_model=JobResponse,
-    tags=["Jobs"],
-    summary="Retry a job",
-    description="Reset a FAILED or DEAD_LETTER job back to PENDING status to retry execution."
-)
-async def retry_job(
-    guid: str,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db)
-):
-    """Resets a FAILED or DEAD_LETTER job to PENDING."""
-    result = await db.execute(select(Job).where(Job.guid == guid))
-    job = result.scalar_one_or_none()
-    if not job:
-        raise HTTPException(status_code=404, detail="Job not found")
-    if job.status not in ("FAILED", "DEAD_LETTER"):
-        raise HTTPException(
-            status_code=409,
-            detail=f"Cannot retry job with status {job.status}. Only FAILED and DEAD_LETTER jobs can be retried."
-        )
-    job.status = "PENDING"
-    job.retry_count = 0
-    job.retry_after = None
-    job.node_id = None
-    job.completed_at = None
-    audit(db, current_user, "job:retry", guid)
-    await db.commit()
-    await ws_manager.broadcast("job:updated", {"guid": guid, "status": "PENDING"})
-    # Construct JobResponse from updated job
-    payload = job.payload if isinstance(job.payload, dict) else json.loads(job.payload or '{}')
-    result_val = job.result if isinstance(job.result, dict) else json.loads(job.result or 'null') if job.result else None
-    target_tags = job.target_tags if isinstance(job.target_tags, list) else json.loads(job.target_tags or 'null') if job.target_tags else None
-    return JobResponse(
-        guid=job.guid,
-        status=job.status,
-        payload=payload,
-        result=result_val,
-        node_id=job.node_id,
-        started_at=job.started_at,
-        duration_seconds=None,
-        target_tags=target_tags,
-        task_type=job.task_type,
-        display_type=getattr(job, 'display_type', None),
-        name=getattr(job, 'name', None),
-        created_by=getattr(job, 'created_by', None),
-        created_at=job.created_at,
-        runtime=getattr(job, 'runtime', None),
-        originating_guid=getattr(job, 'originating_guid', None),
-    )
-
-# --- Resubmit / Bulk job operations (Phase 51) ---
-
-CANCELLABLE_STATES = {"PENDING", "ASSIGNED"}
-RESUBMITTABLE_STATES = {"FAILED", "DEAD_LETTER"}
-TERMINAL_STATES = {"COMPLETED", "FAILED", "DEAD_LETTER", "CANCELLED", "SECURITY_REJECTED"}
-
-
-def _job_to_response(job: Job) -> JobResponse:
-    """Build a JobResponse from a Job ORM object."""
-    payload = json.loads(job.payload) if isinstance(job.payload, str) else job.payload
-    duration = None
-    if job.started_at:
-        end = job.completed_at or datetime.utcnow()
-        duration = (end - job.started_at).total_seconds()
-    return JobResponse(
-        guid=job.guid,
-        status=job.status,
-        payload=payload,
-        result=json.loads(job.result) if job.result else None,
-        node_id=job.node_id,
-        started_at=job.started_at,
-        duration_seconds=duration,
-        target_tags=json.loads(job.target_tags) if job.target_tags else None,
-        depends_on=json.loads(job.depends_on) if job.depends_on else None,
-        task_type=job.task_type,
-        name=job.name,
-        created_by=job.created_by,
-        created_at=job.created_at,
-        runtime=job.runtime,
-        originating_guid=job.originating_guid,
-    )
-
-
-@app.post("/jobs/bulk-cancel", response_model=BulkActionResponse, tags=["Jobs"])
-async def bulk_cancel_jobs(
-    req: BulkJobActionRequest,
-    current_user: User = Depends(require_permission("jobs:write")),
-    db: AsyncSession = Depends(get_db),
-):
-    """Cancel PENDING/ASSIGNED jobs in bulk; skips terminal-state jobs and reports them."""
-    result = await db.execute(select(Job).where(Job.guid.in_(req.guids)))
-    jobs = result.scalars().all()
-    processed, skipped_guids = 0, []
-    for job in jobs:
-        if job.status in CANCELLABLE_STATES:
-            job.status = "CANCELLED"
-            job.completed_at = datetime.utcnow()
-            audit(db, current_user, "job:cancel", job.guid)
-            processed += 1
-        else:
-            skipped_guids.append(job.guid)
-    await db.commit()
-    return BulkActionResponse(processed=processed, skipped=len(skipped_guids), skipped_guids=skipped_guids)
-
-
-@app.post("/jobs/bulk-resubmit", response_model=BulkActionResponse, tags=["Jobs"])
-async def bulk_resubmit_jobs(
-    req: BulkJobActionRequest,
-    current_user: User = Depends(require_permission("jobs:write")),
-    db: AsyncSession = Depends(get_db),
-):
-    """Resubmit FAILED/DEAD_LETTER jobs in bulk; creates a new PENDING job for each."""
-    result = await db.execute(select(Job).where(Job.guid.in_(req.guids)))
-    jobs = result.scalars().all()
-    processed, skipped_guids = 0, []
-    for job in jobs:
-        if job.status in RESUBMITTABLE_STATES:
-            new_guid = str(uuid.uuid4())
-            new_job = Job(
-                guid=new_guid,
-                task_type=job.task_type,
-                payload=job.payload,
-                status="PENDING",
-                target_tags=job.target_tags,
-                capability_requirements=job.capability_requirements,
-                max_retries=job.max_retries,
-                backoff_multiplier=job.backoff_multiplier,
-                timeout_minutes=job.timeout_minutes,
-                runtime=job.runtime,
-                name=job.name,
-                created_by=current_user.username,
-                signature_hmac=job.signature_hmac,
-                originating_guid=job.guid,
-            )
-            db.add(new_job)
-            audit(db, current_user, "job:resubmit", new_guid)
-            processed += 1
-        else:
-            skipped_guids.append(job.guid)
-    await db.commit()
-    return BulkActionResponse(processed=processed, skipped=len(skipped_guids), skipped_guids=skipped_guids)
-
-
-@app.delete("/jobs/bulk", response_model=BulkActionResponse, tags=["Jobs"])
-async def bulk_delete_jobs(
-    req: BulkJobActionRequest,
-    current_user: User = Depends(require_permission("jobs:write")),
-    db: AsyncSession = Depends(get_db),
-):
-    """Delete terminal-state jobs in bulk; skips non-terminal jobs and reports them."""
-    result = await db.execute(select(Job).where(Job.guid.in_(req.guids)))
-    jobs = result.scalars().all()
-    processed, skipped_guids = 0, []
-    for job in jobs:
-        if job.status in TERMINAL_STATES:
-            await db.delete(job)
-            audit(db, current_user, "job:delete", job.guid)
-            processed += 1
-        else:
-            skipped_guids.append(job.guid)
-    await db.commit()
-    return BulkActionResponse(processed=processed, skipped=len(skipped_guids), skipped_guids=skipped_guids)
-
-
-@app.post("/jobs/{guid}/resubmit", response_model=JobResponse, tags=["Jobs"])
-async def resubmit_job(
-    guid: str,
-    current_user: User = Depends(require_permission("jobs:write")),
-    db: AsyncSession = Depends(get_db),
-):
-    """Create a new PENDING job from a FAILED/DEAD_LETTER job, with originating_guid set."""
-    result = await db.execute(select(Job).where(Job.guid == guid))
-    job = result.scalar_one_or_none()
-    if not job:
-        raise HTTPException(status_code=404, detail="Job not found")
-    if job.status not in RESUBMITTABLE_STATES:
-        raise HTTPException(status_code=409, detail="Only FAILED or DEAD_LETTER jobs can be resubmitted")
-    new_guid = str(uuid.uuid4())
-    new_job = Job(
-        guid=new_guid,
-        task_type=job.task_type,
-        payload=job.payload,
-        status="PENDING",
-        target_tags=job.target_tags,
-        capability_requirements=job.capability_requirements,
-        max_retries=job.max_retries,
-        backoff_multiplier=job.backoff_multiplier,
-        timeout_minutes=job.timeout_minutes,
-        runtime=job.runtime,
-        name=job.name,
-        created_by=current_user.username,
-        signature_hmac=job.signature_hmac,
-        originating_guid=guid,
-    )
-    db.add(new_job)
-    audit(db, current_user, "job:resubmit", new_guid)
-    await db.commit()
-    await db.refresh(new_job)
-    await ws_manager.broadcast("job:created", {"guid": new_guid, "status": "PENDING"})
-    return _job_to_response(new_job)
-
-
-@app.post("/work/pull", response_model=PollResponse, tags=["Node Agent"])
-async def pull_work(request: Request, node_id: str = Depends(verify_node_secret), _: str = Depends(verify_client_cert), db: AsyncSession = Depends(get_db)):
-    # LIC-04: DEGRADED_CE — return empty work, nodes stay enrolled and heartbeating
-    _ls = getattr(request.app.state, "licence_state", None)
-    if _ls and _ls.status == LicenceStatus.EXPIRED:
-        from .models import WorkConfig
-        return PollResponse(job=None, config=WorkConfig())
-
-    node_ip = request.client.host
-    r = await db.execute(select(Node).where(Node.node_id == node_id))
-    n = r.scalar_one_or_none()
-    if n and n.status == "REVOKED":
-        raise HTTPException(status_code=403, detail="Node is revoked")
-    return await JobService.pull_work(node_id, node_ip, db)
-
-@app.post(
-    "/heartbeat",
-    response_model=dict,
-    tags=["Node Agent"],
-    summary="Receive node heartbeat",
-    description="Process heartbeat from node agent with health status, resource stats, and system metrics"
-)
-async def receive_heartbeat(req: Request, hb: HeartbeatPayload, node_id: str = Depends(verify_node_secret), _: str = Depends(verify_client_cert), db: AsyncSession = Depends(get_db)):
-    node_ip = req.client.host
-    result = await JobService.receive_heartbeat(node_id, node_ip, hb, db)
-    await ws_manager.broadcast("node:heartbeat", {"node_id": node_id, "status": "ONLINE", "stats": hb.stats})
-    return result
-
-@app.post(
-    "/work/{guid}/result",
-    response_model=dict,
-    tags=["Node Agent"],
-    summary="Report job execution result",
-    description="Node agent reports job completion status, output, and execution metrics"
-)
-async def report_result(guid: str, report: ResultReport, req: Request, node_id: str = Depends(verify_node_secret), db: AsyncSession = Depends(get_db)):
-    node_ip = req.client.host
-    if report.result:
-        report.result = mask_pii(report.result)
-
-    updated = await JobService.report_result(guid, report, node_ip, db)
-    if not updated:
-        raise HTTPException(status_code=404, detail="Job not found")
-
-    # Phase 147: If job is linked to a workflow step, advance the workflow
-    job = await db.get(Job, guid)
-    if job and job.workflow_step_run_id:
-        # Extract run_id from workflow_step_run_id by querying the step run
-        step_run = await db.get(WorkflowStepRun, job.workflow_step_run_id)
-        if step_run:
-            workflow_service = WorkflowService()
-            # NEW: Store result_json for IF gate evaluation
-            if report.result:
-                await workflow_service.store_step_result(step_run.id, report.result, db)
-
-            # Phase 150: Update step_run status based on job completion
-            old_step_status = step_run.status
-            if updated.get("status") == "COMPLETED":
-                step_run.status = "COMPLETED"
-                step_run.completed_at = datetime.utcnow()
-            elif updated.get("status") in ["FAILED", "DEAD_LETTER", "SECURITY_REJECTED"]:
-                step_run.status = "FAILED"
-                step_run.completed_at = datetime.utcnow()
-
-            # Emit workflow_step_updated event if status changed
-            if old_step_status != step_run.status:
-                try:
-                    event = WorkflowStepUpdatedEvent(
-                        id=step_run.id,
-                        workflow_run_id=step_run.workflow_run_id,
-                        workflow_step_id=step_run.workflow_step_id,
-                        status=step_run.status,
-                        started_at=step_run.started_at,
-                        completed_at=step_run.completed_at,
-                        job_guid=guid
-                    )
-                    await ws_manager.broadcast_workflow_step_updated(event)
-                except Exception as e:
-                    logger.error(f"Failed to broadcast workflow_step_updated event: {e}")
-
-            await workflow_service.advance_workflow(step_run.workflow_run_id, db)
-
-    await ws_manager.broadcast("job:updated", {"guid": guid, "status": updated.get("status", "COMPLETED")})
-    return updated
-
-@app.get("/nodes", tags=["Nodes"], response_model=PaginatedResponse[NodeResponse], summary="List all nodes", description="Retrieve paginated list of nodes with online/offline status and capability info")
-async def list_nodes(
-    page: int = 1,
-    page_size: int = 25,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db),
-):
-    # Step 1: total count
-    total_result = await db.execute(select(func.count()).select_from(Node))
-    total = total_result.scalar() or 0
-
-    # Step 2: paginated node list (paginate BEFORE stats batch query)
-    result = await db.execute(
-        select(Node).order_by(Node.hostname)
-        .offset((page - 1) * page_size)
-        .limit(page_size)
-    )
-    nodes = result.scalars().all()
-
-    # Step 3: batch stats history — scoped to this page's node_ids only
-    history_map: dict = defaultdict(list)
-    if nodes:
-        page_node_ids = [n.node_id for n in nodes]
-        hist_result = await db.execute(
-            select(NodeStats)
-            .where(NodeStats.node_id.in_(page_node_ids))
-            .order_by(desc(NodeStats.recorded_at))
-        )
-        for stat in hist_result.scalars().all():
-            bucket = history_map[stat.node_id]
-            if len(bucket) < 20:
-                bucket.append({"t": stat.recorded_at.isoformat(), "cpu": stat.cpu, "ram": stat.ram})
-        # Reverse each bucket so oldest→newest (chronological for charts)
-        for k in history_map:
-            history_map[k].reverse()
-
-    # Step 4: build resp list
-    resp = []
-    for n in nodes:
-        if n.status in ("REVOKED", "TAMPERED", "DRAINING"):
-            node_status = n.status
-        else:
-            is_offline = (datetime.utcnow() - n.last_seen).total_seconds() > 60
-            node_status = "OFFLINE" if is_offline else "ONLINE"
-
-        stats = json.loads(n.stats) if n.stats else None
-        reported_tags = json.loads(n.tags) if n.tags else []
-        op_tags = json.loads(n.operator_tags) if n.operator_tags else None
-        effective_tags = op_tags if op_tags is not None else reported_tags
-
-        resp.append({
-            "node_id": n.node_id,
-            "hostname": n.hostname,
-            "ip": n.ip,
-            "last_seen": n.last_seen,
-            "status": node_status,
-            "base_os_family": n.base_os_family,
-            "stats": stats,
-            "tags": effective_tags,
-            "is_operator_managed": op_tags is not None,
-            "capabilities": json.loads(n.capabilities) if n.capabilities else None,
-            "stats_history": history_map.get(n.node_id, []),
-            "env_tag": n.env_tag,
-            "detected_cgroup_version": n.detected_cgroup_version,
-            "execution_mode": n.execution_mode,  # Phase 124
-        })
-
-    return {
-        "items": resp,
-        "total": total,
-        "page": page,
-        "page_size": page_size,
-    }
-
-@app.get("/nodes/{node_id}/detail", tags=["Nodes"], response_model=NodeResponse, summary="Get node details", description="Retrieve full details of a specific node")
-async def get_node_detail(node_id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    detail = await JobService.get_node_detail(node_id, db)
-    if detail is None:
-        raise HTTPException(status_code=404, detail="Node not found")
-    return detail
-
-
-@app.patch("/nodes/{node_id}", tags=["Nodes"], response_model=ActionResponse, summary="Update node metadata", description="Update node tags and environment tag")
-async def update_node_config(node_id: str, config: NodeUpdateRequest, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Node).where(Node.node_id == node_id))
-    node = result.scalar_one_or_none()
-    if not node:
-        raise HTTPException(status_code=404, detail="Node not found")
-    if config.tags is not None:
-        node.operator_tags = json.dumps(config.tags)
-    if config.env_tag is not None:
-        node.env_tag = config.env_tag if config.env_tag != "" else None
-        node.operator_env_tag = True  # Stays True even when cleared — distinguishes "never touched" from "explicitly cleared"
-
-    await db.commit()
-    return {
-        "status": "updated",
-        "resource_type": "node",
-        "resource_id": node_id,
-    }
-
-@app.delete(
-    "/nodes/{node_id}",
-    status_code=204,
-    response_class=Response,
-    tags=["Nodes"],
-    summary="Delete a node",
-    description="Permanently delete a node and its associated metadata"
-)
-async def delete_node(node_id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Node).where(Node.node_id == node_id))
-    node = result.scalar_one_or_none()
-    if not node:
-        raise HTTPException(status_code=404, detail="Node not found")
-    await db.execute(delete(NodeStats).where(NodeStats.node_id == node_id))
-    await db.execute(delete(Ping).where(Ping.node_id == node_id))
-    audit(db, current_user, "node:delete", node_id)
-    await db.delete(node)
-    await db.commit()
-    return Response(status_code=204)
-
-@app.post("/nodes/{node_id}/revoke", tags=["Nodes"], response_model=ActionResponse, summary="Revoke node certificates", description="Prevent a node from accepting further work and block its enrollment")
-async def revoke_node(node_id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Node).where(Node.node_id == node_id))
-    node = result.scalar_one_or_none()
-    if not node:
-        raise HTTPException(status_code=404, detail="Node not found")
-    if node.status == "REVOKED":
-        raise HTTPException(status_code=409, detail="Node is already revoked")
-    node.status = "REVOKED"
-    if node.client_cert_pem:
-        try:
-            parsed = _x509.load_pem_x509_certificate(node.client_cert_pem.encode())
-            db.add(RevokedCert(serial_number=str(parsed.serial_number), node_id=node_id))
-        except Exception:
-            pass
-    audit(db, current_user, "node:revoke", node_id)
-    await db.commit()
-    return {"status": "revoked", "resource_type": "node", "resource_id": node_id}
-
-@app.patch("/nodes/{node_id}/drain", response_model=ActionResponse, summary="Drain node workload", description="Stop assigning new jobs while completing existing jobs", tags=["Nodes"])
-async def drain_node(node_id: str, current_user: User = Depends(require_permission("nodes:write")), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Node).where(Node.node_id == node_id))
-    node = result.scalar_one_or_none()
-    if not node:
-        raise HTTPException(status_code=404, detail="Node not found")
-    if node.status not in ("ONLINE", "BUSY"):
-        raise HTTPException(status_code=409, detail=f"Cannot drain node in {node.status} state")
-    node.status = "DRAINING"
-    audit(db, current_user, "node:drain", node_id)
-    await db.commit()
-    await ws_manager.broadcast("node:updated", {"node_id": node_id, "status": "DRAINING"})
-    return {"status": "enabled", "resource_type": "node", "resource_id": node_id, "message": "Draining jobs"}
-
-
-@app.patch("/nodes/{node_id}/undrain", response_model=ActionResponse, summary="Resume assigning jobs to drained node", description="Re-enable job assignment on a previously drained node", tags=["Nodes"])
-async def undrain_node(node_id: str, current_user: User = Depends(require_permission("nodes:write")), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Node).where(Node.node_id == node_id))
-    node = result.scalar_one_or_none()
-    if not node:
-        raise HTTPException(status_code=404, detail="Node not found")
-    if node.status != "DRAINING":
-        raise HTTPException(status_code=409, detail="Node is not in DRAINING state")
-    node.status = "ONLINE"
-    audit(db, current_user, "node:undrain", node_id)
-    await db.commit()
-    await ws_manager.broadcast("node:updated", {"node_id": node_id, "status": "ONLINE"})
-    return {"status": "enabled", "resource_type": "node", "resource_id": node_id, "message": "Node accepting jobs"}
-
-
-@app.post("/api/nodes/{node_id}/clear-tamper", response_model=ActionResponse, summary="Clear tamper flag", description="Reset a node from TAMPERED to ONLINE after forensic review", tags=["Nodes"])
-async def clear_node_tamper(node_id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    """Resets a node from TAMPERED to ONLINE after administrator review."""
-    
-    result = await db.execute(select(Node).where(Node.node_id == node_id))
-    node = result.scalar_one_or_none()
-    if not node:
-        raise HTTPException(status_code=404, detail="Node not found")
-    
-    if node.status != "TAMPERED":
-        raise HTTPException(status_code=409, detail="Node is not in tampered state")
-
-    node.status = "ONLINE"
-    node.tamper_details = None
-    await db.commit()
-
-    audit(db, current_user, "node:clear_tamper", node_id)
-    return {"status": "approved", "resource_type": "node", "resource_id": node_id}
-
-@app.post("/nodes/{node_id}/reinstate", response_model=ActionResponse, summary="Reinstate a revoked node", description="Transition a REVOKED node back to OFFLINE status for re-enrollment", tags=["Nodes"])
-async def reinstate_node(node_id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Node).where(Node.node_id == node_id))
-    node = result.scalar_one_or_none()
-    if not node:
-        raise HTTPException(status_code=404, detail="Node not found")
-    if node.status != "REVOKED":
-        raise HTTPException(status_code=409, detail="Node is not revoked")
-    node.status = "OFFLINE"
-    audit(db, current_user, "node:reinstate", node_id)
-    await db.commit()
-    return {"status": "approved", "resource_type": "node", "resource_id": node_id}
-
-@app.post("/auth/register", response_model=RegisterResponse, tags=["Authentication"])
-async def register_node(req: RegisterRequest, db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Token).where(Token.token == req.client_secret))
-    token_entry = result.scalar_one_or_none()
-    
-    if not token_entry:
-         raise HTTPException(status_code=403, detail="Invalid Join Token")
-
-    try:
-        signed_cert = pki_service.sign_csr(req.csr_pem, req.hostname)
-        return {
-            "client_cert_pem": signed_cert,
-            "ca_url": "https://localhost:8001" 
-        }
-    except Exception as e:
-        logger.error(f"CSR Signing Error: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to sign CSR: {str(e)}")
-
-@app.post("/api/enroll", response_model=RegisterResponse, tags=["Node Agent"])
-async def enroll_node(req: EnrollmentRequest, request: Request, db: AsyncSession = Depends(get_db)):
-    """Public endpoint for secure node enrollment using a one-time token."""
-    # LIC-07: Node limit enforcement (checked before token validation to return correct status code)
-    _ls = getattr(request.app.state, "licence_state", None)
-    _node_limit = _ls.node_limit if _ls else 0
-    if _node_limit > 0:
-        from sqlalchemy import text as _sql_text
-        _count_result = await db.execute(
-            _sql_text("SELECT count(*) FROM nodes WHERE status NOT IN ('OFFLINE', 'REVOKED')")
-        )
-        _active_count = _count_result.scalar() or 0
-        if _active_count >= _node_limit:
-            raise HTTPException(
-                status_code=402,
-                detail="Node limit reached — upgrade your licence at axiom.sh/renew"
-            )
-
-    # 1. Verify Token
-    result = await db.execute(select(Token).where(Token.token == req.token, Token.used == False))
-    token_entry = result.scalar_one_or_none()
-
-    if not token_entry:
-         raise HTTPException(status_code=403, detail="Invalid or Expired Enrollment Token")
-
-    # 2. Invalidate Token immediately
-    token_entry.used = True
-
-    try:
-        # 3. Sign CSR
-        signed_cert = pki_service.sign_csr(req.csr_pem, req.hostname)
-
-        # 4. Create or Update Node with the secret binding
-        node_id = req.hostname # Or derived from CSR/Certificate
-        node_ip = request.client.host
-
-        result = await db.execute(select(Node).where(Node.node_id == node_id))
-        node = result.scalar_one_or_none()
-
-        if node:
-            if node.status == "REVOKED":
-                raise HTTPException(status_code=403, detail="Node has been revoked and cannot re-enroll")
-            node.node_secret_hash = req.node_secret_hash
-            node.machine_id = req.machine_id
-            node.ip = node_ip
-            node.last_seen = datetime.utcnow()
-            node.client_cert_pem = signed_cert
-            node.template_id = token_entry.template_id
-        else:
-            node = Node(
-                node_id=node_id,
-                hostname=req.hostname,
-                ip=node_ip,
-                status="ONLINE",
-                template_id=token_entry.template_id,
-                machine_id=req.machine_id,
-                node_secret_hash=req.node_secret_hash,
-                client_cert_pem=signed_cert,
-            )
-            db.add(node)
-        await db.commit()
-        
-        return {
-            "client_cert_pem": signed_cert,
-            "ca_url": f"{request.base_url}" 
-        }
-    except HTTPException:
-        await db.rollback()
-        raise
-    except Exception as e:
-        logger.error(f"Enrollment Error: {e}")
-        await db.rollback()
-        raise HTTPException(status_code=500, detail=f"Enrollment failed: {str(e)}")
-
-# --- Admin Endpoints ---
-
-import base64
-
-@app.post("/admin/generate-token", response_model=EnrollmentTokenResponse, tags=["Admin"])
-@limiter.limit("10/minute")
-async def generate_token(request: Request, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-         
-    token_str = uuid.uuid4().hex
-    token_entry = Token(token=token_str)
-    db.add(token_entry)
-    await db.commit()
-    
-    ca_pem = pki_service.get_root_cert_pem()
-    
-    payload = {
-        "t": token_str,
-        "ca": ca_pem
-    }
-    
-    b64_token = base64.b64encode(json.dumps(payload).encode()).decode()
-    return EnrollmentTokenResponse(token=b64_token)
-
 @app.get(
     "/job-definitions",
     response_model=List[JobDefinitionResponse],
@@ -2267,848 +860,6 @@ async def generate_token(request: Request, current_user: User = Depends(require_
 async def dashboard_job_definitions(current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
     """Dashboard expects /job-definitions instead of /jobs/definitions"""
     return await scheduler_service.list_job_definitions(db)
-
-@app.post("/api/enrollment-tokens", response_model=EnrollmentTokenResponse, tags=["Node Agent"])
-async def create_enrollment_token(req: Optional[EnrollmentTokenCreate] = None, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-
-    token_str = uuid.uuid4().hex
-    token_entry = Token(token=token_str)
-    if req and req.template_id:
-        token_entry.template_id = req.template_id
-
-    db.add(token_entry)
-    await db.commit()
-    return EnrollmentTokenResponse(token=token_str)
-@app.post("/admin/upload-key", response_model=ActionResponse, tags=["Admin"])
-async def upload_public_key(req: UploadKeyRequest, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Config).where(Config.key == "signing_public_key"))
-    row = result.scalar_one_or_none()
-    if row:
-        row.value = req.key_content
-    else:
-        db.add(Config(key="signing_public_key", value=req.key_content))
-    audit(db, current_user, "key:upload")
-    await db.commit()
-    return ActionResponse(status="created", resource_type="public_key", resource_id="signing_public_key", message="Public key uploaded and stored")
-
-# --- Licence Management (Phase 116) ---
-
-@app.post("/api/admin/licence/reload", response_model=LicenceReloadResponse, tags=["Admin"])
-async def reload_licence_endpoint(
-    request: LicenceReloadRequest,
-    current_user: User = Depends(require_permission("system:write")),
-    db: AsyncSession = Depends(get_db)
-):
-    """Hot-reload licence key without restarting the server.
-
-    Args:
-        request: LicenceReloadRequest with optional licence_key override
-        current_user: Must have system:write permission
-        db: Database session
-
-    Returns:
-        LicenceReloadResponse with new licence state
-
-    Raises:
-        HTTPException 422 if licence validation fails
-    """
-    old_state = app.state.licence_state
-
-    try:
-        new_state = await reload_licence(licence_key=request.licence_key)
-    except LicenceError as e:
-        # Invalid licence — keep old state active and return error
-        audit(db, current_user, "licence:reload_failed", detail={"error": str(e), "old_status": old_state.status.value})
-        await db.commit()
-        raise HTTPException(
-            status_code=422,
-            detail={
-                "error": "invalid_licence",
-                "message": str(e)
-            }
-        )
-
-    # Atomic swap — new state is valid
-    app.state.licence_state = new_state
-
-    # Live-activate EE plugins if licence is now valid but plugins aren't loaded
-    ee_activated = False
-    if new_state.is_ee_active and not getattr(app.state.ee, "foundry", False):
-        from .ee import activate_ee_live
-        from .db import engine
-        new_ctx = await activate_ee_live(app, engine)
-        if new_ctx:
-            app.state.ee = new_ctx
-            ee_activated = True
-            logger.info("EE plugins live-activated via licence reload")
-
-    # Broadcast licence status change to all connected WebSocket clients
-    await ws_manager.broadcast("licence_status_changed", {
-        "old_status": old_state.status.value,
-        "new_status": new_state.status.value,
-        "message": f"Licence updated to {new_state.status.value}",
-        "timestamp": datetime.utcnow().isoformat(),
-        "metadata": {
-            "organization": new_state.customer_id or "Unknown",
-            "tier": new_state.tier,
-            "expires_at": (datetime.utcnow() + timedelta(days=new_state.days_until_expiry)).isoformat() if new_state.days_until_expiry > 0 else None,
-            "ee_activated": ee_activated,
-        }
-    })
-
-    # If EE was just activated, broadcast feature flags so dashboards update
-    if ee_activated:
-        ctx = app.state.ee
-        await ws_manager.broadcast("features_changed", {
-            "foundry": ctx.foundry, "audit": ctx.audit, "webhooks": ctx.webhooks,
-            "triggers": ctx.triggers, "rbac": ctx.rbac, "resource_limits": ctx.resource_limits,
-            "service_principals": ctx.service_principals, "api_keys": ctx.api_keys,
-            "executions": ctx.executions,
-        })
-
-    # Audit the transition
-    audit(
-        db, current_user, "licence:reload_success",
-        detail={
-            "old_status": old_state.status.value,
-            "new_status": new_state.status.value,
-            "tier": new_state.tier,
-            "customer_id": new_state.customer_id,
-            "node_limit": new_state.node_limit,
-            "days_until_expiry": new_state.days_until_expiry,
-            "ee_activated": ee_activated,
-        }
-    )
-    await db.commit()
-
-    return LicenceReloadResponse(
-        status=new_state.status.value,
-        tier=new_state.tier,
-        customer_id=new_state.customer_id,
-        node_limit=new_state.node_limit,
-        grace_days=new_state.grace_days,
-        days_until_expiry=new_state.days_until_expiry,
-        features=new_state.features,
-        is_ee_active=new_state.is_ee_active
-    )
-
-@app.get(
-    "/config/public-key",
-    response_model=dict,
-    tags=["System"],
-    summary="Get signing public key",
-    description="Returns the Ed25519 public key for job script signature verification (requires valid enrollment token)"
-)
-async def get_public_key(x_join_token: str = Header(None), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(Token).where(Token.token == x_join_token))
-    if not result.scalar_one_or_none():
-         raise HTTPException(status_code=403, detail="Invalid Join Token")
-
-    result = await db.execute(select(Config).where(Config.key == "signing_public_key"))
-    row = result.scalar_one_or_none()
-    if not row:
-        raise HTTPException(status_code=404, detail="Key not found")
-    return {"public_key": row.value}
-
-# --- Configuration Endpoints ---
-
-@app.get("/config/mounts", response_model=List[NetworkMount], tags=["System"])
-async def get_network_mounts(
-    db: AsyncSession = Depends(get_db), 
-    user: Optional[User] = Depends(get_current_user_optional),
-    x_join_token: Optional[str] = Header(None)
-):
-    is_authed = user is not None
-    is_valid_token = False
-
-    if x_join_token:
-         result = await db.execute(select(Token).where(Token.token == x_join_token))
-         if result.scalar_one_or_none():
-             is_valid_token = True
-
-    if not (is_authed or is_valid_token):
-         raise HTTPException(status_code=403, detail="Not Authorized")
-
-    result = await db.execute(select(Config).where(Config.key == "global_network_mounts"))
-    row = result.scalar_one_or_none()
-    if not row:
-        return []
-    try:
-        data = json.loads(row.value)
-        return [NetworkMount(**m) for m in data]
-    except:
-        return []
-
-@app.post("/config/mounts", response_model=ActionResponse, tags=["System"])
-async def update_network_mounts(config: MountsConfig, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-
-    for m in config.mounts:
-        if not m.name.replace("_", "").isalnum():
-             raise HTTPException(status_code=400, detail=f"Invalid mount name: {m.name}")
-
-    json_str = json.dumps([m.dict() for m in config.mounts])
-
-    result = await db.execute(select(Config).where(Config.key == "global_network_mounts"))
-    row = result.scalar_one_or_none()
-    if row:
-        row.value = json_str
-    else:
-        db.add(Config(key="global_network_mounts", value=json_str))
-
-    await db.commit()
-    return {"status": "updated", "resource_type": "mounts", "resource_id": "global_network_mounts", "message": f"Updated {len(config.mounts)} mount(s)"}
-
-# --- Signature Registry API ---
-
-@app.post("/signatures", response_model=SignatureResponse, tags=["Signatures"])
-async def upload_signature(sig: SignatureCreate, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    return await SignatureService.upload_signature(sig, current_user, db)
-
-@app.get("/signatures", response_model=List[SignatureResponse], tags=["Signatures"])
-async def list_signatures(current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    return await SignatureService.list_signatures(db)
-
-@app.get("/signatures/{id}", response_model=SignatureResponse, tags=["Signatures"])
-async def get_signature(id: str, db: AsyncSession = Depends(get_db)):
-    """Get a signature by ID. Unauthenticated (nodes need to fetch this for verification)."""
-    result = await db.execute(select(Signature).where(Signature.id == id))
-    sig = result.scalar_one_or_none()
-    if not sig:
-        raise HTTPException(status_code=404, detail="Signature not found")
-    return sig
-
-@app.delete("/signatures/{id}", response_model=ActionResponse, tags=["Signatures"])
-async def delete_signature(id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    success = await SignatureService.delete_signature(id, db)
-    if not success:
-        raise HTTPException(status_code=404, detail="Signature not found")
-    audit(db, current_user, "signature:delete", id)
-    await db.commit()
-    return {"status": "deleted", "resource_type": "signature", "resource_id": id}
-
-# --- Job Definitions API ---
-
-@app.post("/jobs/definitions", response_model=JobDefinitionResponse, tags=["Job Definitions"])
-async def create_job_definition(def_req: JobDefinitionCreate, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    return await scheduler_service.create_job_definition(def_req, current_user, db)
-
-@app.get("/jobs/definitions", response_model=List[JobDefinitionResponse], tags=["Job Definitions"])
-async def list_job_definitions(current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    return await scheduler_service.list_job_definitions(db)
-
-@app.delete("/jobs/definitions/{id}", response_model=ActionResponse, tags=["Job Definitions"])
-async def delete_job_definition(id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(ScheduledJob).where(ScheduledJob.id == id))
-    job_def = result.scalar_one_or_none()
-    if not job_def:
-        raise HTTPException(status_code=404, detail="Job definition not found")
-    try:
-        scheduler_service.scheduler.remove_job(id)
-    except Exception:
-        pass
-    await db.delete(job_def)
-    await db.commit()
-    return {"status": "deleted", "resource_type": "job_definition", "resource_id": id}
-
-@app.patch("/jobs/definitions/{id}/toggle", response_model=ActionResponse, tags=["Job Definitions"])
-async def toggle_job_definition(id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(ScheduledJob).where(ScheduledJob.id == id))
-    job_def = result.scalar_one_or_none()
-    if not job_def:
-        raise HTTPException(status_code=404, detail="Job definition not found")
-    job_def.is_active = not job_def.is_active
-    await db.commit()
-    await scheduler_service.sync_scheduler()
-    return {"status": "updated", "resource_type": "job_definition", "resource_id": id, "message": f"Job definition is now {'active' if job_def.is_active else 'inactive'}"}
-
-@app.get("/jobs/definitions/{id}", response_model=JobDefinitionResponse, tags=["Job Definitions"])
-async def get_job_definition(id: str, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    return await scheduler_service.get_job_definition(id, db)
-
-@app.get("/api/schedule", response_model=ScheduleListResponse, tags=["Schedule"])
-async def get_schedule(
-    db: AsyncSession = Depends(get_db),
-    _: User = Depends(require_permission("jobs:read"))
-) -> ScheduleListResponse:
-    """
-    Unified schedule view: merges ScheduledJob and cron-scheduled Workflow entries.
-    Returns sorted by next_run_time ascending (soonest first).
-    Only includes active items with cron schedules.
-    """
-    return await scheduler_service.get_unified_schedule(db)
-
-
-@app.post("/api/jobs/push", response_model=JobDefinitionResponse, status_code=201, tags=["Job Definitions"])
-async def push_job_definition(
-    req: JobPushRequest,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db),
-):
-    """RFC-compliant push endpoint: creates DRAFT or updates existing job with dual JWT+Ed25519 verification."""
-    # 1. Validate Ed25519 signature BEFORE any DB write (STAGE-03)
-    sig_result = await db.execute(select(Signature).where(Signature.id == req.signature_id))
-    sig = sig_result.scalar_one_or_none()
-    if not sig:
-        raise HTTPException(404, detail="Signature ID not found")
-    try:
-        SignatureService.verify_payload_signature(sig.public_key, req.signature, req.script_content)
-    except Exception as e:
-        raise HTTPException(422, detail=(
-            "Signature verification failed — the script content does not match the provided signature. "
-            "Ensure you signed the exact script content with the private key paired to the registered public key. "
-            "See the Signatures page in the dashboard for key generation instructions."
-        ))
-
-    # 2. Identity attribution (STAGE-04)
-    pushed_by = current_user.username  # "username" or "sp:name" for service principals
-
-    # 3. Upsert logic (STAGE-02)
-    if req.id:
-        # Update existing job by ID
-        result = await db.execute(select(ScheduledJob).where(ScheduledJob.id == req.id))
-        job = result.scalar_one_or_none()
-        if not job:
-            raise HTTPException(404, detail="Job definition not found")
-        if job.status == "REVOKED":
-            raise HTTPException(409, detail={"error": "job_revoked", "id": job.id,
-                                             "message": "Job is REVOKED. Un-REVOKE to DEPRECATED before re-pushing."})
-        job.script_content = req.script_content
-        job.signature_id = req.signature_id
-        job.signature_payload = req.signature
-        job.pushed_by = pushed_by
-        job.updated_at = datetime.utcnow()
-    else:
-        # Create new job by name — check for name conflict first
-        existing_result = await db.execute(select(ScheduledJob).where(ScheduledJob.name == req.name))
-        existing = existing_result.scalar_one_or_none()
-        if existing:
-            raise HTTPException(409, detail={"error": "name_conflict", "id": existing.id,
-                                             "message": f"Job '{req.name}' already exists. Use id to update."})
-        job = ScheduledJob(
-            id=uuid.uuid4().hex,
-            name=req.name,
-            script_content=req.script_content,
-            signature_id=req.signature_id,
-            signature_payload=req.signature,
-            schedule_cron="",  # DRAFT jobs have no schedule yet
-            status="DRAFT",
-            pushed_by=pushed_by,
-            created_by=pushed_by,
-        )
-        db.add(job)
-
-    audit(db, current_user, "job:pushed", job.id if req.id else None,
-          {"name": req.name or job.name, "pushed_by": pushed_by, "action": "update" if req.id else "create"})
-    await db.commit()
-    await db.refresh(job)
-    return JobDefinitionResponse.model_validate(job)
-
-@app.patch("/jobs/definitions/{id}", response_model=JobDefinitionResponse, tags=["Job Definitions"])
-async def update_job_definition(id: str, update_req: JobDefinitionUpdate, current_user: User = Depends(require_auth), db: AsyncSession = Depends(get_db)):
-    return await scheduler_service.update_job_definition(id, update_req, current_user, db)
-
-# === Workflow Routes ===
-
-@app.post("/api/workflows", tags=["workflows"], response_model=WorkflowResponse, status_code=201)
-async def create_workflow(
-    workflow_create: WorkflowCreate,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-) -> WorkflowResponse:
-    """Create a new Workflow with full-graph contract."""
-    workflow_service = WorkflowService()
-    return await workflow_service.create(db, workflow_create, current_user.username)
-
-@app.get("/api/workflows", tags=["workflows"], response_model=list[WorkflowResponse])
-async def list_workflows(
-    skip: int = Query(0, ge=0),
-    limit: int = Query(100, ge=1, le=1000),
-    db: AsyncSession = Depends(get_db)
-) -> list[WorkflowResponse]:
-    """List all Workflows with metadata (step_count, last_run_status)."""
-    workflow_service = WorkflowService()
-    return await workflow_service.list(db, skip, limit)
-
-@app.get("/api/workflows/{workflow_id}", tags=["workflows"], response_model=WorkflowResponse)
-async def get_workflow(
-    workflow_id: str,
-    db: AsyncSession = Depends(get_db)
-) -> WorkflowResponse:
-    """Get a single Workflow with full DAG (nested steps, edges, parameters)."""
-    workflow_service = WorkflowService()
-    return await workflow_service.get(db, workflow_id)
-
-@app.put("/api/workflows/{workflow_id}", tags=["workflows"], response_model=WorkflowResponse)
-async def update_workflow(
-    workflow_id: str,
-    workflow_update: WorkflowUpdate,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-) -> WorkflowResponse:
-    """Update a Workflow definition (atomic replace of steps/edges/parameters)."""
-    workflow_service = WorkflowService()
-    return await workflow_service.update(db, workflow_id, workflow_update)
-
-@app.patch("/api/workflows/{workflow_id}", tags=["workflows"], response_model=WorkflowResponse)
-async def patch_workflow(
-    workflow_id: str,
-    workflow_update: WorkflowUpdate,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-) -> WorkflowResponse:
-    """
-    Partially update a Workflow definition with validation.
-
-    Validates schedule_cron against required parameters (all must have defaults).
-    Updates DAG if definition changes. Syncs APScheduler crons after save.
-
-    Returns:
-        200 OK with updated WorkflowResponse
-
-    Raises:
-        404: Workflow not found
-        422: Validation failed (e.g., schedule_cron with required params lacking defaults)
-    """
-    # Fetch workflow
-    workflow = await db.get(Workflow, workflow_id)
-    if workflow is None:
-        raise HTTPException(status_code=404, detail="Workflow not found")
-
-    # If schedule_cron is being set, validate that all required parameters have defaults
-    if workflow_update.schedule_cron is not None:
-        # Fetch parameters for this workflow
-        stmt = select(Workflow).where(Workflow.id == workflow_id).options(
-            selectinload(Workflow.parameters)
-        )
-        result = await db.execute(stmt)
-        workflow_with_params = result.scalar_one_or_none()
-
-        if workflow_with_params and workflow_with_params.parameters:
-            for param in workflow_with_params.parameters:
-                if param.default_value is None:
-                    raise HTTPException(
-                        status_code=422,
-                        detail=f"Cannot schedule workflow: required parameter '{param.name}' has no default value. All parameters must have defaults for cron scheduling."
-                    )
-
-    # Update fields
-    if workflow_update.name is not None:
-        workflow.name = workflow_update.name
-    if workflow_update.definition is not None:
-        workflow.definition = workflow_update.definition
-        # Validate DAG if definition changed
-        workflow_service = WorkflowService()
-        await workflow_service.validate_workflow_definition(workflow)
-    if workflow_update.schedule_cron is not None:
-        workflow.schedule_cron = workflow_update.schedule_cron
-    if workflow_update.is_paused is not None:
-        workflow.is_paused = workflow_update.is_paused
-
-    workflow.updated_at = datetime.now(UTC)
-    await db.commit()
-
-    # Audit log
-    audit(db, current_user, "update_workflow", workflow_id, {
-        "name": workflow_update.name,
-        "schedule_cron": workflow_update.schedule_cron,
-        "is_paused": workflow_update.is_paused
-    })
-
-    # Sync crons with APScheduler if schedule_cron changed
-    try:
-        await scheduler_service.sync_workflow_crons()
-    except Exception as e:
-        logger.warning(f"Error syncing workflow crons: {e}")
-
-    return await workflow_service._workflow_to_response(db, workflow)
-
-@app.delete("/api/workflows/{workflow_id}", tags=["workflows"], status_code=204)
-async def delete_workflow(
-    workflow_id: str,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-) -> None:
-    """Delete a Workflow (blocked if active WorkflowRuns exist)."""
-    workflow_service = WorkflowService()
-    await workflow_service.delete(db, workflow_id)
-
-@app.post("/api/workflows/{workflow_id}/fork", tags=["workflows"], response_model=WorkflowResponse, status_code=201)
-async def fork_workflow(
-    workflow_id: str,
-    fork_request: dict = Body({"new_name": "..."}),  # {"new_name": "cloned-workflow"}
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-) -> WorkflowResponse:
-    """
-    Fork (Save-as-New) a Workflow: clone all steps/edges/parameters into a new Workflow,
-    and pause the source workflow's cron schedule.
-    """
-    new_name = fork_request.get("new_name")
-    if not new_name:
-        raise HTTPException(status_code=422, detail="new_name is required")
-
-    workflow_service = WorkflowService()
-    return await workflow_service.fork(db, workflow_id, new_name, current_user.username)
-
-@app.post("/api/workflows/validate", tags=["workflows"], response_model=dict)
-async def validate_workflow(
-    workflow_create: WorkflowCreate
-) -> dict:
-    """
-    Validate a Workflow definition without saving (static check).
-    Used by the DAG editor (Phase 151) on every canvas change.
-    Returns {valid: true} or {valid: false, error: ..., cycle_path: ..., etc.}
-    """
-    steps_data = [s.model_dump() for s in workflow_create.steps]
-    edges_data = [e.model_dump() for e in workflow_create.edges]
-
-    # Generate temporary IDs for validation (like create() does)
-    for i, step in enumerate(steps_data):
-        if "id" not in step:
-            step["id"] = f"temp_step_{i}"
-
-    is_valid, error = WorkflowService.validate_dag(steps_data, edges_data)
-    if is_valid:
-        return {"valid": True}
-    else:
-        return {"valid": False, **error}
-
-@app.post("/api/workflow-runs", tags=["workflows"], response_model=WorkflowRunResponse, status_code=201)
-async def create_workflow_run(
-    body: dict,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-):
-    """
-    Manually trigger a WorkflowRun with optional parameter overrides.
-
-    Request body: {
-        "workflow_id": str,
-        "parameters": {key: value} (optional)
-    }
-
-    Returns WorkflowRunResponse with initial step runs.
-
-    Returns:
-        201 Created with WorkflowRunResponse
-
-    Raises:
-        400: workflow_id missing
-        404: Workflow not found
-        409: Workflow is paused
-        422: Required parameters unsatisfied
-        500: Unexpected error
-    """
-    if not body.get("workflow_id"):
-        raise HTTPException(status_code=400, detail="workflow_id required")
-
-    workflow_service = WorkflowService()
-    try:
-        run = await workflow_service.start_run(
-            workflow_id=body["workflow_id"],
-            parameters=body.get("parameters", {}),
-            trigger_type="MANUAL",
-            triggered_by=current_user.username,
-            db=db
-        )
-
-        # Audit log the manual trigger
-        audit(db, current_user, "trigger_workflow_manual", run.id, {
-            "workflow_id": body["workflow_id"],
-            "run_id": run.id,
-            "triggered_by": current_user.username
-        })
-
-        await ws_manager.broadcast("workflow:run:created", {"run_id": run.id, "workflow_id": run.workflow_id, "status": "RUNNING"})
-        return run
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Error triggering workflow {body.get('workflow_id')}: {e}")
-        raise HTTPException(status_code=500, detail="Failed to trigger workflow")
-
-@app.post("/api/workflows/{workflow_id}/webhooks", tags=["workflows"], response_model=WorkflowWebhookResponse, status_code=201)
-async def create_workflow_webhook(
-    workflow_id: str,
-    webhook_create: WorkflowWebhookCreate,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-) -> WorkflowWebhookResponse:
-    """Create a webhook endpoint for a Workflow. Returns plaintext secret once at creation; never again."""
-    # Verify workflow exists
-    workflow = await db.get(Workflow, workflow_id)
-    if workflow is None:
-        raise HTTPException(status_code=404, detail="Workflow not found")
-
-    # Generate webhook ID and plaintext secret
-    webhook_id = str(uuid.uuid4())
-    plaintext_secret = _secrets.token_urlsafe(32)  # ~256 bits of entropy
-
-    # Hash secret for storage (bcrypt)
-    secret_hash = hash_webhook_secret(plaintext_secret)
-
-    # Encrypt plaintext secret for HMAC verification at trigger time (Fernet)
-    secret_plaintext_encrypted = cipher_suite.encrypt(plaintext_secret.encode()).decode()
-
-    # Create webhook in DB
-    webhook = WorkflowWebhook(
-        id=webhook_id,
-        workflow_id=workflow_id,
-        name=webhook_create.name,
-        secret_hash=secret_hash,
-        secret_plaintext=secret_plaintext_encrypted,  # Store encrypted plaintext for HMAC verification
-        created_at=datetime.now(UTC)
-    )
-    db.add(webhook)
-    await db.commit()
-
-    # Audit log
-    audit(db, current_user, "create_webhook", webhook_id, {
-        "workflow_id": workflow_id,
-        "webhook_name": webhook_create.name
-    })
-
-    # Return response with plaintext secret (only time it's exposed)
-    response = WorkflowWebhookResponse(
-        id=webhook_id,
-        workflow_id=workflow_id,
-        name=webhook_create.name,
-        secret=plaintext_secret,  # ONLY in creation response
-        created_at=webhook.created_at
-    )
-    return response
-
-@app.get("/api/workflows/{workflow_id}/webhooks", tags=["workflows"], response_model=List[WorkflowWebhookResponse])
-async def list_workflow_webhooks(
-    workflow_id: str,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-) -> List[WorkflowWebhookResponse]:
-    """List all webhooks for a Workflow (secret field omitted for security)."""
-    # Verify workflow exists
-    workflow = await db.get(Workflow, workflow_id)
-    if workflow is None:
-        raise HTTPException(status_code=404, detail="Workflow not found")
-
-    # Fetch all webhooks for this workflow
-    stmt = select(WorkflowWebhook).where(WorkflowWebhook.workflow_id == workflow_id)
-    result = await db.execute(stmt)
-    webhooks = result.scalars().all()
-
-    # Map to response (secret=None for all list responses)
-    return [
-        WorkflowWebhookResponse(
-            id=w.id,
-            workflow_id=w.workflow_id,
-            name=w.name,
-            secret=None,  # Never expose secret in list
-            created_at=w.created_at
-        )
-        for w in webhooks
-    ]
-
-@app.delete("/api/workflows/{workflow_id}/webhooks/{webhook_id}", tags=["workflows"], status_code=204)
-async def delete_workflow_webhook(
-    workflow_id: str,
-    webhook_id: str,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-) -> None:
-    """Revoke a webhook endpoint. Future trigger requests will fail with 404."""
-    # Verify webhook exists and belongs to this workflow
-    webhook = await db.get(WorkflowWebhook, webhook_id)
-    if webhook is None or webhook.workflow_id != workflow_id:
-        raise HTTPException(status_code=404, detail="Webhook not found")
-
-    # Delete webhook
-    await db.delete(webhook)
-    await db.commit()
-
-    # Audit log
-    audit(db, current_user, "delete_webhook", webhook_id, {
-        "workflow_id": workflow_id
-    })
-
-@app.get("/api/workflows/{workflow_id}/runs", tags=["workflows"], response_model=WorkflowRunListResponse)
-async def get_workflow_runs(
-    workflow_id: str,
-    skip: int = Query(0, ge=0),
-    limit: int = Query(100, ge=1, le=1000),
-    current_user: User = Depends(require_permission("workflows:read")),
-    db: AsyncSession = Depends(get_db)
-) -> WorkflowRunListResponse:
-    """Get paginated list of runs for a workflow (most recent first)."""
-    # Verify workflow exists
-    workflow = await db.get(Workflow, workflow_id)
-    if workflow is None:
-        raise HTTPException(status_code=404, detail="Workflow not found")
-
-    # Count total runs
-    count_stmt = select(func.count(WorkflowRun.id)).where(
-        WorkflowRun.workflow_id == workflow_id
-    )
-    count_result = await db.execute(count_stmt)
-    total = count_result.scalar() or 0
-
-    # Fetch paginated runs (order by started_at DESC, most recent first)
-    stmt = (
-        select(WorkflowRun)
-        .where(WorkflowRun.workflow_id == workflow_id)
-        .order_by(desc(WorkflowRun.started_at))
-        .offset(skip)
-        .limit(limit)
-    )
-    runs_result = await db.execute(stmt)
-    runs = runs_result.scalars().all()
-
-    # Convert to response models (with eager-loaded step runs)
-    from .services.workflow_service import WorkflowService
-    workflow_service = WorkflowService()
-    run_responses = []
-    for run in runs:
-        run_response = await workflow_service._run_to_response(db, run)
-        run_responses.append(run_response)
-
-    return WorkflowRunListResponse(
-        runs=run_responses,
-        total=total,
-        skip=skip,
-        limit=limit,
-    )
-
-@app.get("/api/workflows/{workflow_id}/runs/{run_id}", tags=["workflows"], response_model=WorkflowRunResponse)
-async def get_workflow_run(
-    workflow_id: str,
-    run_id: str,
-    current_user: User = Depends(require_permission("workflows:read")),
-    db: AsyncSession = Depends(get_db),
-) -> WorkflowRunResponse:
-    """Get a single WorkflowRun with all step details."""
-    run = await db.get(WorkflowRun, run_id)
-    if run is None or run.workflow_id != workflow_id:
-        raise HTTPException(status_code=404, detail="Run not found")
-    from .services.workflow_service import WorkflowService
-    return await WorkflowService()._run_to_response(db, run)
-
-
-@app.post("/api/workflow-runs/{run_id}/cancel", tags=["workflows"], response_model=WorkflowRunResponse)
-async def cancel_workflow_run(
-    run_id: str,
-    current_user: User = Depends(require_permission("workflows:write")),
-    db: AsyncSession = Depends(get_db)
-):
-    """
-    Cancel a running WorkflowRun.
-
-    Blocks new step dispatches; running jobs continue to completion.
-    """
-    workflow_service = WorkflowService()
-    run = await workflow_service.cancel_run(run_id, db)
-
-    await ws_manager.broadcast("workflow:run:cancelled", {"run_id": run.id, "status": "CANCELLED"})
-    return run
-
-@app.post("/api/webhooks/{webhook_id}/trigger", tags=["webhooks"], status_code=202)
-async def trigger_webhook(
-    webhook_id: str,
-    request: Request,
-    db: AsyncSession = Depends(get_db)
-) -> Dict[str, str]:
-    """
-    Unauthenticated webhook trigger endpoint. Validates HMAC signature before running workflow.
-
-    Request body: JSON object with parameter overrides (e.g., {"env": "prod", "region": "us-east-1"})
-
-    Header: X-Hub-Signature-256: sha256=<hex> (HMAC-SHA256 of raw request body)
-
-    Returns:
-        202 Accepted with {"run_id": "<workflow_run_id>"}
-
-    Raises:
-        400: Invalid request body
-        401: Missing or invalid signature
-        404: Webhook not found
-        500: Failed to trigger workflow
-    """
-    # Read raw request body (required for HMAC verification)
-    try:
-        body_bytes = await request.body()
-    except Exception as e:
-        logger.error(f"Error reading webhook body: {e}")
-        raise HTTPException(status_code=400, detail="Invalid request body")
-
-    # Extract signature header
-    signature_header = request.headers.get("X-Hub-Signature-256", "")
-    if not signature_header:
-        logger.warning(f"Webhook {webhook_id} missing X-Hub-Signature-256 header")
-        raise HTTPException(status_code=401, detail="Missing signature header")
-
-    # Fetch webhook
-    webhook = await db.get(WorkflowWebhook, webhook_id)
-    if webhook is None:
-        # Log but don't expose webhook existence
-        logger.warning(f"Webhook trigger attempt for unknown webhook {webhook_id}")
-        raise HTTPException(status_code=404, detail="Webhook not found")
-
-    # Get plaintext secret from webhook for verification
-    # The secret_plaintext is stored Fernet-encrypted; decrypt it for HMAC verification
-    if webhook.secret_plaintext is None:
-        logger.error(f"Webhook {webhook_id} has no plaintext secret stored")
-        raise HTTPException(status_code=500, detail="Webhook misconfiguration")
-
-    try:
-        plaintext_secret = cipher_suite.decrypt(webhook.secret_plaintext.encode()).decode()
-    except Exception as e:
-        logger.error(f"Webhook {webhook_id} decryption failed: {e}")
-        raise HTTPException(status_code=500, detail="Webhook decryption failed")
-
-    # Verify signature
-    if not verify_webhook_signature(signature_header, body_bytes, plaintext_secret):
-        logger.warning(f"Webhook {webhook_id} signature verification failed")
-        raise HTTPException(status_code=401, detail="Signature verification failed")
-
-    # Parse request body as parameters dict
-    try:
-        parameters = await request.json()  # Should be JSON object
-        if not isinstance(parameters, dict):
-            parameters = {}
-    except Exception as e:
-        logger.warning(f"Webhook {webhook_id} invalid JSON body: {e}")
-        parameters = {}  # Treat as empty if JSON invalid
-
-    # Trigger workflow run
-    try:
-        workflow_service = WorkflowService()
-        run = await workflow_service.start_run(
-            workflow_id=webhook.workflow_id,
-            parameters=parameters,
-            trigger_type="WEBHOOK",
-            triggered_by=webhook.name,
-            db=db
-        )
-
-        # Audit log (no user context for unauthenticated endpoint)
-        try:
-            from sqlalchemy import text
-            await db.execute(
-                text("INSERT INTO audit_log (username, action, resource_id, detail) VALUES (:u, :a, :r, :d)"),
-                {"u": "webhook", "a": "trigger_webhook", "r": run.id, "d": json.dumps({
-                    "webhook_id": webhook_id,
-                    "workflow_id": webhook.workflow_id,
-                    "run_id": run.id
-                })}
-            )
-        except Exception:
-            # Audit table may not exist in CE mode — silently ignore
-            pass
-
-        await ws_manager.broadcast("workflow:run:created", {"run_id": run.id, "workflow_id": run.workflow_id, "status": "RUNNING"})
-        return {"run_id": run.id}  # 202 + run_id
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Error triggering webhook {webhook_id}: {e}")
-        raise HTTPException(status_code=500, detail="Failed to trigger workflow")
 
 # --- Installer & Doc Endpoints ---
 
@@ -3185,143 +936,6 @@ async def get_doc_content(filename: str, current_user: User = Depends(require_au
     with open(safe_path, "r") as f:
         content = f.read()
     return {"content": content}
-
-# --- WebSocket Live Feed ---
-
-@app.websocket("/ws")
-async def websocket_endpoint(ws: WebSocket, token: Optional[str] = None):
-    """Live event feed. Requires a valid JWT passed as ?token=<jwt> query param."""
-    await ws.accept()
-    # Validate token using a short-lived session so we don't hold a pool slot
-    # for the entire WebSocket lifetime (which exhausts the connection pool).
-    authed = False
-    if token:
-        try:
-            from jose import jwt as _jwt, JWTError
-            from .auth import SECRET_KEY, ALGORITHM
-            payload = _jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-            username = payload.get("sub")
-            if username:
-                async with AsyncSessionLocal() as _db:
-                    result = await _db.execute(select(User).where(User.username == username))
-                    user = result.scalar_one_or_none()
-                    if user and payload.get("tv", 0) == user.token_version:
-                        authed = True
-        except Exception:
-            pass
-    if not authed:
-        await ws.close(code=1008)
-        return
-    ws_manager._connections.append(ws)
-    try:
-        while True:
-            data = await ws.receive_text()
-            if data == "ping":
-                await ws.send_text("pong")
-    except WebSocketDisconnect:
-        ws_manager.disconnect(ws)
-
-# --- CRL Endpoint ---
-
-@app.get(
-    "/system/crl.pem",
-    response_class=Response,
-    tags=["System"],
-    summary="Get Certificate Revocation List",
-    description="Returns signed X.509 CRL (Certificate Revocation List) of all revoked node client certificates"
-)
-async def get_crl(db: AsyncSession = Depends(get_db)):
-    """Returns a signed X.509 CRL of all revoked node certificates."""
-    result = await db.execute(select(RevokedCert))
-    revoked = result.scalars().all()
-    serials = [r.serial_number for r in revoked]
-    crl_pem = pki_service.ca_authority.generate_crl(serials)
-    return Response(content=crl_pem, media_type="application/x-pem-file")
-
-# --- Signal API (Reactive Orchestration) ---
-
-@app.post(
-    "/api/signals/{name}",
-    response_model=SignalResponse,
-    tags=["Headless Automation"],
-    summary="Fire a named signal",
-    description="Fire a named signal to unblock dependent jobs waiting on signal conditions"
-)
-async def fire_signal(
-    name: str,
-    req: Optional[SignalFire] = None,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db)
-):
-    """
-    Fire a named signal to unblock dependent jobs.
-
-    Authenticated via Bearer Token.
-    """
-
-    # Upsert signal
-    result = await db.execute(select(Signal).where(Signal.name == name))
-    sig = result.scalar_one_or_none()
-    
-    payload_json = json.dumps(req.payload) if req and req.payload else None
-    
-    if sig:
-        sig.payload = payload_json
-        sig.created_at = datetime.utcnow()
-    else:
-        sig = Signal(name=name, payload=payload_json)
-        db.add(sig)
-    
-    audit(db, current_user, "signal:fire", name)
-    await db.commit()
-    
-    # Trigger workflow advancement for SIGNAL_WAIT steps
-    await workflow_service.advance_signal_wait(name, db)
-    
-    # Trigger unblocking
-    await JobService.unblock_jobs_by_signal(name, db)
-    
-    return {"status": "fired", "name": name}
-
-@app.get("/api/signals", response_model=List[SignalResponse], tags=["Headless Automation"])
-async def list_signals(
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db)
-):
-    """List all currently active signals (Admin Only)."""
-    result = await db.execute(select(Signal).order_by(Signal.created_at.desc()))
-    signals = result.scalars().all()
-    
-    resp = []
-    for s in signals:
-        resp.append(SignalResponse(
-            name=s.name,
-            payload=json.loads(s.payload) if s.payload else None,
-            created_at=s.created_at
-        ))
-    return resp
-
-@app.delete(
-    "/api/signals/{name}",
-    response_model=dict,
-    tags=["Headless Automation"],
-    summary="Clear a signal",
-    description="Delete a signal from the system"
-)
-async def clear_signal(
-    name: str,
-    current_user: User = Depends(require_auth),
-    db: AsyncSession = Depends(get_db)
-):
-    """Clear a signal from the system (Admin Only)."""
-    result = await db.execute(select(Signal).where(Signal.name == name))
-    sig = result.scalar_one_or_none()
-    if not sig:
-        raise HTTPException(404, "Signal not found")
-    
-    await db.delete(sig)
-    await db.commit()
-    return {"status": "cleared"}
 
 # --- Job Templates (SRCH-06, SRCH-07) ---
